% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extrule.R
\name{extrule}
\alias{extrule}
\title{Bound of rules}
\usage{
extrule(myrules, varname)
}
\arguments{
\item{myrules}{A 3 column matrix output of function "hughs.path.rpart"}

\item{varname}{the names of x variables}
}
\value{
A p*2 matrix, p is the length of varname. The first column is the lower bound, the second column is the upper bound. The default lower bound is "-Inf",the default upper bound is "Inf". row corresponse to x variables ordered in the data matrix given to rpart.
}
\description{
This function takes in a ruleset and output the lower and upper bounds of each rule.
}

