% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harfunc.R
\name{harfunc}
\alias{harfunc}
\title{A harvested classification tree}
\usage{
harfunc(rpart.object, data, varname, sig = 0.95)
}
\arguments{
\item{rpart.object}{classification result of training data from traditional classification tree(rpart function).}

\item{data}{original training data where 'y' stores classmembership}

\item{varname}{the name of each explaanatory variables}

\item{sig}{significance level (default 0.95)}
}
\value{
the list of orginial result of classification, likelihood improvment and harvested classification result.
}
\description{
Basic function to apply the harvest algorithm to the training data set, computing whether we can harvest any nodes based on the classic classification tree algorithm.
}

