% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.harfunc.R
\name{predict}
\alias{predict}
\title{Predictions from a harvested tree}
\usage{
predict(harfunc.object, data, num.var)
}
\arguments{
\item{harfunc.object}{the output of harfunc function}

\item{data}{test data}

\item{num.var}{number of explaining varibles}
}
\value{
"pred.mat" the result of prediction, belong is the node that data point belongs to, possibility is the probability of point being in class 1, predict is the simple perdict based on whether probability is larger than 0.5.
}
\description{
Return a predicted membership from a harvested tree
}

