% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rank.nodes.R
\name{rank.nodes}
\alias{rank.nodes}
\title{Ranking of nodes}
\usage{
rank.nodes(harfunc.object)
}
\arguments{
\item{harfunc.object}{an object of class "harfunc"}
}
\value{
the ranked harvest nodes
}
\description{
Rank harvested node by lower p value
}

