% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rulesets.R
\name{rulesets}
\alias{rulesets}
\title{A logical matrix for a terminal node}
\usage{
rulesets(noden, newsim, varn, nodenumb)
}
\arguments{
\item{noden}{a terminal node defined by a set of rules, from function "treemat"}

\item{newsim}{data to be harvested}

\item{varn}{x variable names}

\item{nodenumb}{all the labels of terminal nodes}
}
\value{
A nxnn logical matrix, n=number of data points to be harvested, nn=number of rules defining a terminal node. Each column of the matrix corresponding to a node that is defined by one variable/rule, its name corresponds to that variable. Note the original terminal node is just the intersection of these nodes.
}
\description{
Return a logical matrix of the rule sets which define a terminal node
}

