\name{training}
\alias{training}
\docType{data}
\title{
training
}
\description{
A simulated data set of symptoms of breast cancer patients
}
\usage{data(training)}
\format{
  A data frame with 300 observations on the following 5 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x1}}{a factor with levels \code{21} \code{22} \code{23} \code{24} \code{25} \code{26} \code{27} \code{28} \code{29}}
    \item{\code{x2}}{a factor with levels \code{39-} \code{40-49} \code{50-69} \code{70-74} \code{75+}}
    \item{\code{x3}}{a numeric vector}
    \item{\code{x4}}{a factor with levels \code{2004} \code{2005} \code{2006} \code{2007} \code{2008} \code{2009} \code{2010}}
  }
}
\source{
simulated data for breast cancer diagnosis
}


\keyword{datasets}
