% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest}
\alias{harvest}
\title{A harvested classification tree}
\usage{
harvest(training, num.var, numeric.info, sig = 0.95)
}
\arguments{
\item{training}{original data where 'y' stores classmembership 0 and 1,in the first column, with explanatory variable stores in the second to the last column.}

\item{num.var}{number of explanatory variables}

\item{numeric.info}{the vector stores the number of which variable is continuous}

\item{sig}{significance level (default 0.95)}
}
\value{
the original set, the likelihood improvment and the harvested classfication rules
}
\description{
The main function of the package, aiming at develop the harvest classification tree. Training data input and
}
\examples{
data(training)
harvest(training,4,NA)
}

