\name{Harvest.Tree}
\alias{Harvest.Tree}
\alias{Harvest.Tree}
\docType{package}
\title{
Harvest the classification tree
}
\description{
This package is aimed at applying the Harvest classification tree algorithm, which is a modified algorithm of classic classification tree. It was firstly used in drug discovery field, but it also performs well in other kind of data, especially when active region is unrelated.

To learn more about the harvest classification algorithm, you can go to http://www.stat.ubc.ca/Research/TechReports/techreports/220.pdf for more imformation.
}
\details{
\tabular{ll}{
Package: \tab Harvest.Tree\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-09\cr
License: \tab GPL-2\cr
}
The main function of package calls 'harfunc', it can be used to analysis the data which is classified by rpart function. You can try harvest() function first to figure out how to use the package. The
'predict' funciton offers function to predict the unclassified data based on training model.
}
\author{
Bingyuan Liu \\
Yan Yuan \\
Qian Shi


Maintainer: Bingyuan Liu<adler1016@gmail.com>

}

